+++ flake.nix
@@ -53,6 +53,7 @@
             || builtins.match ".*\\.md$" path != null
             || builtins.match ".*\\.toml$" path != null
             || builtins.match ".*\\.json$" path != null
+            || builtins.match ".*\\.jsonl$" path != null
             || builtins.match ".*\\.pkl$" path != null
             || builtins.match ".*\\.html$" path != null
             || builtins.match ".*\\.js$" path != null